+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3]
+### Fixed
+- Fixed `toHex` method used by `sha256` in `Hashing` which had a bug where single
+  digit hex values were not being left-padded with a zero.
+
 ## [0.4.2] - 2022-03-22
 ### Added
 - `makeUniqueSeq` creates unique names by adding a sequential prefix (instead of